third_party/blink/web_tests/fast/scrolling/inertial-scrolling-with-pointer-events-none-overlay.html

<!DOCTYPE HTML>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-actions.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<script src="../../resources/gesture-util.js"></script>

<style>
#container {
  width: 400px;
  height: 400px;
  overflow: auto;
}

#inner {
  height: 3000px;
  background-color: #eee;
}

#overlay {
  position: absolute;
  left: 0;
  top: 0;
  width: 100%;
  height: 100%;
  pointer-events: none;
}

p {
 margin: 0;
 padding: 1000px 0;
}
</style>

<body style="margin:0" onload=runTest()>
  <div id="container">
    <div id="inner"></div>    
  </div>
  <div id="overlay"></div>
</body>

<script>
const container = document.getElementById('container');
const inner = document.getElementById('inner');

const update = () => inner.innerHTML = '<p>Content</p>';
setInterval(update, 200);

function runTest() {
  promise_test (async (t) => {
    const pixels_to_scroll = 100;
    const start_x = 200;
    const start_y = 200;
    const speed_in_pixels_s = 900;

    // Inertial touch scroll not supported on a Mac.
    const isMac = navigator.platform.toUpperCase().indexOf('MAC')>=0;
    if (isMac)
      return;

    await waitForCompositorCommit();
    await gestureScroll(() => {
      return swipe(pixels_to_scroll, start_x, start_y, 'up', speed_in_pixels_s);
    }, container);
    assert_greater_than(container.scrollTop, pixels_to_scroll, 
      "container should scroll at least 100 pixels, which is the length of the swipe.");

    const scroll_top_previous_value = container.scrollTop;

    await waitForCompositorCommit();
    await gestureScroll(() => {
      return swipe(pixels_to_scroll, start_x, start_y, 'up', speed_in_pixels_s);
    }, container);
    assert_greater_than(container.scrollTop, scroll_top_previous_value + pixels_to_scroll);
  }, "Make sure inertial scrolling is not broken with pointer-events:none overlay.");
}
</script>
